requirements.txt

flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
werkzeug==2.3.7


python-dotenv==1.0.0


pytest==7.4.0
pytest-cov==4.1.0


flask-debugtoolbar==0.13.1
